// web/src/models/DTOs/Cobrancas/ParcelaDTO.ts
import { SituacaoCobrancaCliente } from "./SituacaoCobrancaCliente";
import { StatusInterface } from "./StatusInterface";

export interface ParcelaDTO {
  idInterface: number;
  contId: number;
  idSistema: number;
  statusContrato: StatusInterface;
  integraProtheus: string | null;
  sacadoNome: string;
  sacadoCpfCnpj: string;
  contratoProtheus: string | null;
  quantidadeDeParcelas: number;
  idCobrancaCliente: number;
  idCancelamento: string | null;
  dataGeracao: string;
  valorCobranca: number;
  valorAgente: number;
  valorBolsa: number;
  valorDescontoComercial: number;
  valorPago: number;
  dataBaixa: string | null;
  numeroParcela: number;
  dataVencimento: string;
  dataPagamento: string;
  dataCredito: string | null;
  formaPagamento: string | null;
  multa: number;
  juros: number;
  imposto: number;
  retirarJuros: string | null;
  motivoRetiradaJuros: string | null;
  situacao: SituacaoCobrancaCliente;
  estornar: boolean;
  dataEstorno: string | null;
  valorEstorno: number;
  idCobrancaPagamento: number;
  idUnidade: number;
  codigoUnidade: string;
  dataEmissaoNotaFiscal: string | null;
  dataInclusao: string;
  dataAlteracao: string;
  idOperadorInclusao: number;
  idOperadorAlteracao: number | null;
  porcentagemEstornar: number;
  origemValorEstornado: string | null;
  dataAlteracaoProtheus: string;
  dataEstornoProtheus: string | null;
  dataInclusaoProtheus: string | null;
  descontoAteVencimento: number;
  recno: number | null;
  entidade: string;
  formaPagamentoSimplificado: string | null;
}